pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/v3i"]

[project]
name = "v3i"
version = "0.0.1"
description = "VIII"
readme = "README.md"
authors = [
    { name = "Hieronim Kubica", email = "hkkubica@gmail.com" },
]
requires-python = ">=3.12"
dependencies = [
    "numpy==2.2.*",
    "numpy-quaternion==2024.0.*",
    "torch==2.6.*",
    "torchvision==0.21.*",
    "matplotlib==3.10.*",
    "scikit-learn==1.6.*",
    "tqdm==4.67.*",
    "streamlit==1.46.*",
    "plotly==6.1.*",
    "watchdog==6.0.*",
    "PyYAML==6.0.*",
]

[tool.uv]
dev-dependencies = [
    "pip>=25.0.0",
    "pytest>=8.0.0",
    "ruff>=0.12.0",
]
